replaced concodium lib in favor lokal package, to properly integrate SDK (#272)
