+++ Dependencies/MobileWalletFacade.swift
@@ -1,5 +1,4 @@
 import Foundation
-import MobileWallet
 
 
 enum WalletError: Error {
